pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "px-device-interfaces"
version = "0.0.1"
description = "PX Device Interfaces: connection and GPIO helper libraries"
readme = "README.md"
requires-python = ">=3.8"
authors = [ { name = "Yannik Dolde" } ]
license = "MIT"
dependencies = [
    "pyserial",
    "opcua",
    "cryptography",
    "dataclasses-json"
]

[tool.setuptools.packages.find]
include = ["px_device_interfaces*"]
